+++ src/string/firstOf_found.js
@@ -1,0 +1,6 @@
+this.size = l[j].length - firstSize
+this.idx = j
+res = i - offset
+// Reduce the scope of the pattern search
+s = s.substr(offset, res)
+offset = 0