+++ test/rules-methods-trimLeft_no_trimRight-test.js
@@ -218,6 +218,21 @@ describe('Tokenizer Rules Methods with trimRight disabled', function () {
           .write('abcbac')
         })
       })
+
+      describe('#addRule(["a","b"], ["c","d"])', function () {
+        var i = 0
+        it('should return b then a', function (done) {
+          p.addRule(['a','b'], ['c','d'], function (token, idx, type) {
+            i == 0 && assert.equal(token, 'bc')
+            i == 1 && assert.equal(token, 'ac')
+            i == 2 && assert.equal(token, 'bd')
+            i == 3 && assert.equal(token, 'ad')
+            i++
+            if (i == 4) done()
+          })
+          .write('abcbacabdbad')
+        })
+      })
     })
 /*
 ['ab','cd']
@@ -308,6 +323,57 @@ describe('Tokenizer Rules Methods with trimRight disabled', function () {
             if (i == 4) done()
           })
           .write('abcdbcabdabcebcabe')
+        })
+      })
+    })
+/*
+[fn1, fn2]
+[fn1, fn2], 'a'
+[fn1, fn2], ['a','b']
+**/
+    describe('First: Array Second: "..."', function () {
+      function fn1 (s, start) {
+        return s[start] === 'a' ? 1 : -1
+      }
+      function fn2 (s, start) {
+        return s[start] === 'b' ? 1 : -1
+      }
+
+      describe('#addRule([fn1,fn2])', function () {
+        var i = 0
+        it('should return an empty token', function (done) {
+          p.addRule([fn1,fn2], function (token, idx, type) {
+            assert.equal(token, '')
+            i++
+            if (i == 2) done()
+          })
+          .write('ab')
+        })
+      })
+
+      describe('#addRule([fn1,fn2], "a")', function () {
+        var i = 0
+        it('should return bc twice', function (done) {
+          p.addRule([fn1,fn2], 'a', function (token, idx, type) {
+            i == 0 && assert.equal(token, 'bca')
+            i == 1 && assert.equal(token, 'bca')
+            i++
+            if (i == 2) done()
+          })
+          .write('abcabbca')
+        })
+      })
+
+      describe('#addRule([fn1,fn2], ["a","b"])', function () {
+        var i = 0
+        it('should return c 4 times', function (done) {
+          p.addRule([fn1,fn2], ['a','b'], function (token, idx, type) {
+            i == 0 && assert.equal(token, 'ca')
+            i > 0 && assert.equal(token, 'cb')
+            i++
+            if (i == 4) done()
+          })
+          .write('acaacbbcbbcb')
         })
       })
     })